requirements.txt
flask
pandas
numpy
cohere
annoy
python-dotenv
umap-learn